requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
plotly==5.17.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pandas==2.0.3
numpy==2.0.3
scikit-learn==1.3.0
joblib==1.3.2
python-dotenv==1.0.0
pytest==7.3.1
pytest-cov==4.0.0
alembic==1.7.7
pytest-asyncio==0.20.3
pytest-mock==3.7.0
pytest-cov==4.0.0
pytest-asyncio==0.20.3
pytest-mock==3.7.0
pytest-cov==4.0.0
pytest-asyncio==0.20.3
pytest-mock==3.7.0
pytest-cov==4.0.0
pytest-asyncio==0.20.
pytest-mock==3.7.0
pytest-cov==4.0.0
pytest-asyncio==0.20.3
pytest-mock==3.7.0
sckit-learn==1.3.0
